Rename project to axle
It's shorter and easier to pronounce.

PR: https://github.com/whalbawi/axle/pull/7
