+++ CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 3.20)
-project(legatus CXX)
+project(axle CXX)
 
 set(CMAKE_CXX_STANDARD 20)
 
@@ -7,12 +7,12 @@ enable_testing()
 
 
 # Directories
-set(LEGATUS_SRC_DIR ${CMAKE_CURRENT_SOURCE_DIR}/src)
-set(LEGATUS_INCLUDE_DIR ${CMAKE_CURRENT_SOURCE_DIR}/include)
-set(LEGATUS_TEST_DIR ${CMAKE_CURRENT_SOURCE_DIR}/test)
+set(AXLE_SRC_DIR ${CMAKE_CURRENT_SOURCE_DIR}/src)
+set(AXLE_INCLUDE_DIR ${CMAKE_CURRENT_SOURCE_DIR}/include)
+set(AXLE_TEST_DIR ${CMAKE_CURRENT_SOURCE_DIR}/test)
 
-set(LEGATUS_THIRD_PARTY_DIR ${CMAKE_CURRENT_SOURCE_DIR}/third_party)
-set(LEGATUS_GOOGLETEST_DIR ${LEGATUS_THIRD_PARTY_DIR}/googletest)
+set(AXLE_THIRD_PARTY_DIR ${CMAKE_CURRENT_SOURCE_DIR}/third_party)
+set(AXLE_GOOGLETEST_DIR ${AXLE_THIRD_PARTY_DIR}/googletest)
 
 add_compile_options(
     -Wall
@@ -49,55 +49,55 @@ if(ENABLE_UBSAN)
 endif()
 
 # Dependencies
-add_subdirectory(${LEGATUS_GOOGLETEST_DIR} EXCLUDE_FROM_ALL)
+add_subdirectory(${AXLE_GOOGLETEST_DIR} EXCLUDE_FROM_ALL)
 
 set(CMAKE_EXPORT_COMPILE_COMMANDS ON)
 
 # Project version
-file(READ VERSION LEGATUS_PROJECT_VERSION)
-string(STRIP ${LEGATUS_PROJECT_VERSION} LEGATUS_PROJECT_VERSION)
+file(READ VERSION AXLE_PROJECT_VERSION)
+string(STRIP ${AXLE_PROJECT_VERSION} AXLE_PROJECT_VERSION)
 
 # and commit
 execute_process(COMMAND git describe --always --dirty
-    OUTPUT_VARIABLE LEGATUS_GIT_COMMIT_HASH
-    ERROR_VARIABLE LEGATUS_GIT_COMMIT_ERROR
+    OUTPUT_VARIABLE AXLE_GIT_COMMIT_HASH
+    ERROR_VARIABLE AXLE_GIT_COMMIT_ERROR
     OUTPUT_STRIP_TRAILING_WHITESPACE
-    RESULT_VARIABLE LEGATUS_GIT_COMMIT_RESULT
+    RESULT_VARIABLE AXLE_GIT_COMMIT_RESULT
 )
-if (NOT ${LEGATUS_GIT_COMMIT_RESULT} STREQUAL 0)
-    message(WARNING "Could not retrieve git commit: " ${LEGATUS_GIT_COMMIT_ERROR})
-    set(LEGATUS_GIT_COMMIT_HASH unknown)
+if (NOT ${AXLE_GIT_COMMIT_RESULT} STREQUAL 0)
+    message(WARNING "Could not retrieve git commit: " ${AXLE_GIT_COMMIT_ERROR})
+    set(AXLE_GIT_COMMIT_HASH unknown)
 endif()
 
-message("legatus version information: " ${LEGATUS_PROJECT_VERSION} " - " ${LEGATUS_GIT_COMMIT_HASH})
-configure_file(${LEGATUS_SRC_DIR}/version.cpp.in ${CMAKE_CURRENT_BINARY_DIR}/gen/version.cpp @ONLY)
+message("axle version information: " ${AXLE_PROJECT_VERSION} " - " ${AXLE_GIT_COMMIT_HASH})
+configure_file(${AXLE_SRC_DIR}/version.cpp.in ${CMAKE_CURRENT_BINARY_DIR}/gen/version.cpp @ONLY)
 
 # Source files
-set(LEGATUS_SRC_LIST
-    ${LEGATUS_SRC_DIR}/legatus.cpp
-    ${LEGATUS_SRC_DIR}/io/event.cpp
-    ${LEGATUS_SRC_DIR}/io/socket.cpp
+set(AXLE_SRC_LIST
+    ${AXLE_SRC_DIR}/axle.cpp
+    ${AXLE_SRC_DIR}/event.cpp
+    ${AXLE_SRC_DIR}/socket.cpp
     ${CMAKE_CURRENT_BINARY_DIR}/gen/version.cpp
 )
 
 # Test files
-set(LEGATUS_TEST_LIST
-    ${LEGATUS_TEST_DIR}/legatus_test.cpp
-    ${LEGATUS_TEST_DIR}/status_test.cpp
-    ${LEGATUS_TEST_DIR}/io/event_test.cpp
+set(AXLE_TEST_LIST
+    ${AXLE_TEST_DIR}/axle_test.cpp
+    ${AXLE_TEST_DIR}/event_test.cpp
+    ${AXLE_TEST_DIR}/status_test.cpp
 )
 
-add_library(legatus-lib ${LEGATUS_SRC_LIST})
-target_include_directories(legatus-lib PUBLIC ${LEGATUS_INCLUDE_DIR} ${LEGATUS_SRC_DIR})
-set_target_properties(legatus-lib PROPERTIES OUTPUT_NAME legatus)
+add_library(axle-lib ${AXLE_SRC_LIST})
+target_include_directories(axle-lib PUBLIC ${AXLE_INCLUDE_DIR} ${AXLE_SRC_DIR})
+set_target_properties(axle-lib PROPERTIES OUTPUT_NAME axle)
 
-add_executable(legatus-tests ${LEGATUS_TEST_LIST})
-target_include_directories(legatus-tests PRIVATE ${LEGATUS_SRC_DIR})
-target_link_libraries(legatus-tests legatus-lib gtest_main)
+add_executable(axle-tests ${AXLE_TEST_LIST})
+target_include_directories(axle-tests PRIVATE ${AXLE_SRC_DIR})
+target_link_libraries(axle-tests axle-lib gtest_main)
 
-add_test(unit-tests legatus-tests)
+add_test(unit-tests axle-tests)
 
-file(GLOB_RECURSE HDR_FILES "${LEGATUS_SRC_DIR}/*.h" "${LEGATUS_INCLUDE_DIR}/*.h")
+file(GLOB_RECURSE HDR_FILES "${AXLE_SRC_DIR}/*.h" "${AXLE_INCLUDE_DIR}/*.h")
 add_custom_target(lint
   COMMAND /usr/local/bin/clang-tidy -p ${CMAKE_BINARY_DIR} --config-file ${CMAKE_CURRENT_SOURCE_DIR}/.clang-tidy ${HDR_FILES}
   COMMAND ${CMAKE_CURRENT_SOURCE_DIR}/tools/run-clang-tidy.py -p=${CMAKE_BINARY_DIR} -config-file=${CMAKE_CURRENT_SOURCE_DIR}/.clang-tidy
